changed Accouts table schema, updated packages

+++ src/Lykke.Service.Balances/Models/ClientBalances/ClientBalanceResponceModel.cs
@@ -6,8 +6,8 @@ namespace Lykke.Service.Balances.Models.ClientBalances
     public class ClientBalanceResponseModel
     {
         public string AssetId { get; set; }
-        public double Balance { get; set; }
-        public double Reserved { get; set; }
+        public decimal Balance { get; set; }
+        public decimal Reserved { get; set; }
 
         public static ClientBalanceResponseModel Create(IWallet src)
         {
